requirements.txt
pysdl2-dll
pysdl2